fix(ecs): reduce ecs service task role cloudwatch permissions when no log configured (under feature flag) (#31475)
### Issue # (if applicable)

Closes https://github.com/aws/aws-cdk/issues/31397

### Reason for this change

If a FargateService or EC2 Service has `enableExecuteCommand: true` and the ECS cluster it runs on has `executeCommandConfiguration.logging` set to anything but `ecs.ExecuteCommandLogging.NONE` then the CDK automatically grants the underlying TaskDefinition overly broad cloudwatch logs permissions regardless of need even if the logging configuration has no cloudwatch logs config set.

This is not right behaviour as these permissions are not needed. 

### Description of changes

Add feature flag and if feature flag is on and cloudwatch log is configured, reduce the permissions. 

### Description of how you validated changes

New unit tests, integration tests.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*