chore: introduce `promiseall-no-unbounded-parallelism` linter rule
Since JavaScript is single-threaded, `Promise.all()` will only ever be used for
I/O-bound tasks. However, I/O-parallelism isn't free either. Every async I/O-performing
task launched will consume some FDs, and their amount is limited. If the amount
of promises launched is a function of the input the program runs on, the system
FDs might be exhausted. Some concurrency limit must be introduced.

This linter rule exists to remind the programmer of that fact. It triggers
on every `Promise.all()` invocation and cannot be resolved; the only solution
is to think about it, and then silence this rule as proof that you thought about it.