feat(cdk): expose authorizer id and authorization type (#31622)
### Issue # (if applicable)

Closes #31605.

### Reason for this change

I need to access the authorizer id and type to be able to import the authorizer in another CDK project

### Description of changes

I have added a public readonly variable for the authorizationType that was previously a string in the return object from the bind function. 
I have also added a getter that will return the authorizerId after the first binding. If not bound yet it will throw an error.

### Description of how you validated changes

I have added unit tests for all changed files.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*