fix(elasticloadbalancingv2): `http2Enabled` with true is ignored in ApplicationLoadBalancer (#31675)
### Issue # (if applicable)

Closes #31609.

### Reason for this change



The `http2Enabled` can only be set to CFn if it is specified with false. Therefore, if it is specified with true, the change is not reflected in CFn.

### Description of changes



Passed `http2Enabled` if it is not undefined.

```ts
if (props.http2Enabled !== undefined) { this.setAttribute('routing.http2.enabled', props.http2Enabled ? 'true' : 'false'); }
```

### Description of how you validated changes



Both unit and integ tests.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*