fix(cli): `cdk import` errors with 'S3 error: Access Denied'
In #31597 we changed `cdk diff` to always use the file asset publishing
role, instead of direct CLI credentials. This included a refactor that
impacted `cdk import`, which was now not uploading the stack template
at all anymore. The operation that is now broken only happens in a
case with interactive input, which is why this wasn't caught by integ
tests.

In this change, put the requisite asset-handling code around
`makeBodyParameter` to make the asset uploading happen properly.

In future PRs:

- Add an integration test for `cdk import` which would have exposed the
  same error.
- Refactor the contract of `makeBodyParameter`, and perhaps more around
  asset uploading, to make the expectations and promises of that
  function more clear; right now it was not obvious what the function
  would and wouldn't do for you, which led to this error.