Q generated with some of @moelasmar feedback.
Prompt= "I do not think this L2 is following CDK coding guidelines. It missed some properties. Also, for constructs defined in cdk-lib, we do not import from the dependencies, but instead we depend on the modules defined in the project. We shuld use enum where it make sense, as it is a mandatory property. E.g. "engine" property. Given this, can you update the code for elasticache ?"